Migrate JNI to jni_generator for class CobaltTextToSpeechHelper (#4710)
This PR is a syntax migration for the CobaltTextToSpeechHelper class to
unblock the larger migration of StarboardBridge, which depends on
CobaltTextToSpeechHelper.

Parts of this class, or even the entire class, might no longer be
necessary. We will need to investigate and clean it up further under bug
b/392178584.

No tests have been conducted because the class is currently not wired up
in Chrobalt.

b/389117681
original-hexsha: 82896864420b1158e87754b42de454c798b88bcd