Implement custom API for HTMLVideoElement (#5648)
Implements the maxVideoCapabilities API for HTMLVideoElement. This
allows platforms to properly allocate resources for secondary videos,
which are assumed to be at much lower resolutions than the primary
video.

setMaxVideoCapabilities can enable the platform to optimize the playback
pipeline for a low quality video stream if it knows that it will never
adapt to a higher quality stream. The setMaxVideoCapabilities function
must be called before the src attribute is set on element, otherwise an
InvalidStateError exception is raised.

The design document and example use cases for the
setMaxVideoCapabilities function can be seen
[here](https://docs.google.com/document/d/1LJNh1LrBhfblop9Xpevg0Z7yMBJtZX8eD1g-Nn1YjEA/edit?tab=t.0#heading=h.u0p1e6iom4ri).

b/326827007
original-hexsha: 6c593e01072b8102ad1e6897c329280f7b1d94de