Add non-blocking-write compliance test for pipe2 (#6044)
This new test could be single-threaded and very simple but would hang on
platforms that are noncompliant and block on writes to full pipes that
are configured to be non-blocking. To improve the experience for these
platforms, a second thread fails the test after a timeout by performing
a large read on the pipe to unblock the write.

Issue: 412690906
original-hexsha: 142e601cd351280e45381930b00e7875687f3cd8